Host the `Retropad_360pad.png` ourself
The image linked in the GitHub wiki [1] and then also in the /doc/
section [2] has moved. We could (a) just find out the new url on
libretro.com, or (b) we host the image ourself. I prefer (b).

I have also fixed the url in wiki, hence the update of the commit
for /doc/ section.

[1] https://github.com/libretro/Lakka-LibreELEC/wiki/Input-settings
[2] https://www.lakka.tv/doc/input-settings/